Fixing the stories (comments) not appearing
The root of the problem is that we were creating a copy of the task when showing the editor. However, that copy was no longer reactive.

Rather than make it reactive (which is just adding work and causing issues with backwards references), the fix here is to remove the "in place updating" that we were doing on the task, and instead push the inputs to bind to individual fields. At save, we then apply those field values to the task.