Choose the most recent version of project.
Previously, even though we were fetching all projects on refresh, we were retaining older versions of the project. This is because we were placing newer versions at the end of the projects array. Hence, when we use unique(), it prefers the older version (since it comes first). This has been addressed by placing newer projects at the front of the list.